requirements.txt
fastapi
uvicorn
python-multipart
openai
jinja2
fpdf2
python-dotenv

